requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
marshmallow==3.20.1
flask-restx==1.2.0
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0